% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundary_points.R
\name{get_boundary_points}
\alias{get_boundary_points}
\title{get_boundary_points}
\usage{
get_boundary_points(X)
}
\arguments{
\item{X}{a numeric matrix with one point per row.}
}
\value{
A numeric matrix of the same number of columns as \code{X}.
}
\description{
Given a set of points, returns the minimal subset with the same upper hull.
}
\examples{
X <- rbind(c(0, 1), c(1, 0), c(1, 0), c(1, 1))
get_boundary_points(X)

}
